Extract userSettings handling into SettingsClient
