+++ packages/apps/desktop-viewer-test/src/frontend/components/home/Home.tsx
@@ -39,11 +39,14 @@ const Home = () => {
     });
   }, []);
 
+  // Get latest recent user settings on mount
+  // to make sure that the deleted files are being checked and handled correctly
   useEffect(() => {
     const getUserSettings = async () => {
       await userSettings.getUserSettings();
     };
     void getUserSettings();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => {
